requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
python-dotenv
flask-cors
werkzeug
requests
pyttsx3
SpeechRecognition
openai-whisper
pymongo
pyaudio 